[topgen] Import SVD generation support into topgen.py.
This mostly duplicates the contents of svdtool.py, except that topgen
uses different (and more accurate) semantics for piecing together the
IP module definitions.

As a result, while the two generate similar output, the contents are
not exactly equivalent. svdtool.py's days are likely numbered...
