+++ python-requirements.txt
@@ -15,6 +15,7 @@ meson >= 0.51.0, < 0.53.0 # minimum matches version in meson.build
 mistletoe>=0.7.2
 pyftdi
 pygments
+pysvd
 pytest
 pytest-timeout
 pyyaml
